refactoring searcher interface
